platform/do: move retry logic to platform
The api package wraps the error from DO so the shouldRetry logic doesn't
detect the DO errors correctly. Move the logic to the api package so it
can correctly detect the error type.
